pyproject.toml
[tool.poetry]
name = "counselhub"
version = "0.1.0"
description = "Django counsel hub"
authors = [
    "Ali minuei <aliminuei@gmail.com>"
]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11"
django = ">=5.2,<6.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-decouple = ">=3.8,<4.0"
django-simple-history = "^3.8.0"
django-health-check = "^3.18.3"
django-jalali-date = "^2.0.0"
django-import-export = { version = "^4.3.7", extras = ["xlsx"] }
django-cors-headers = "^4.7.0"
django-leaflet = ">=0.31.0,<0.32"
jdatetime = "^5.2.0"
black = "^25.1.0"
gunicorn = "23.0.0"
djangorestframework-simplejwt = "^5.3.0"
pyjwt = "^2.8.0"
drf-spectacular = "0.26.2"
persian = ">=0.5.0,<0.6"
persiantools = ">=3.0.1"
django-admin-interface = ">=0.30.1,<0.31"
httpx = ">=0.25.0,<0.26"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.25.0"

[tool.poe.tasks]
compilemessages = { shell = "python3 manage.py compilemessages -l fa" }
makemessages = { shell = "python3 manage.py makemessages -l fa" }